requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
chromadb
faiss-cpu
pillow
poppler
pdf2image